Corrections for 128-bit link

+++ src/pke/lib/scheme/bfvrns/bfvrns-cryptoparameters.cpp
@@ -71,7 +71,8 @@ void CryptoParametersBFVRNS::PrecomputeCRTTables(KeySwitchTechnique ksTech, Scal
     m_modqBarrettMu.resize(sizeQ);
     for (uint32_t i = 0; i < sizeQ; i++) {
         BigInteger mu = BarrettBase128Bit / BigInteger(moduliQ[i]);
-        uint64_t val[2]{(mu % TwoPower64).ConvertToInt(), mu.RShift(64).ConvertToInt()};
+        uint64_t val[2]{static_cast<uint64_t>((mu % TwoPower64).ConvertToInt()),
+                        static_cast<uint64_t>(mu.RShift(64).ConvertToInt())};
         memcpy(&m_modqBarrettMu[i], val, sizeof(DoubleNativeInt));
     }
 
@@ -141,7 +142,8 @@ void CryptoParametersBFVRNS::PrecomputeCRTTables(KeySwitchTechnique ksTech, Scal
         m_modrBarrettMu.resize(sizeR);
         for (uint32_t i = 0; i < sizeR; i++) {
             BigInteger mu = BarrettBase128Bit / BigInteger(moduliR[i]);
-            uint64_t val[2]{(mu % TwoPower64).ConvertToInt(), mu.RShift(64).ConvertToInt()};
+            uint64_t val[2]{static_cast<uint64_t>((mu % TwoPower64).ConvertToInt()),
+                            static_cast<uint64_t>(mu.RShift(64).ConvertToInt())};
             memcpy(&m_modrBarrettMu[i], val, sizeof(DoubleNativeInt));
         }
 
@@ -256,7 +258,8 @@ void CryptoParametersBFVRNS::PrecomputeCRTTables(KeySwitchTechnique ksTech, Scal
         m_modrBarrettMu.resize(sizeR);
         for (uint32_t j = 0; j < moduliR.size(); j++) {
             BigInteger mu = BarrettBase128Bit / BigInteger(moduliR[j]);
-            uint64_t val[2]{(mu % TwoPower64).ConvertToInt(), mu.RShift(64).ConvertToInt()};
+            uint64_t val[2]{static_cast<uint64_t>((mu % TwoPower64).ConvertToInt()),
+                            static_cast<uint64_t>(mu.RShift(64).ConvertToInt())};
             memcpy(&m_modrBarrettMu[j], val, sizeof(DoubleNativeInt));
         }
 
@@ -714,7 +717,8 @@ void CryptoParametersBFVRNS::PrecomputeCRTTables(KeySwitchTechnique ksTech, Scal
         m_modbskBarrettMu.resize(m_moduliBsk.size());
         for (uint32_t i = 0; i < m_modbskBarrettMu.size(); i++) {
             BigInteger mu = BarrettBase128Bit / BigInteger(m_moduliBsk[i]);
-            uint64_t val[2]{(mu % TwoPower64).ConvertToInt(), mu.RShift(64).ConvertToInt()};
+            uint64_t val[2]{static_cast<uint64_t>((mu % TwoPower64).ConvertToInt()),
+                            static_cast<uint64_t>(mu.RShift(64).ConvertToInt())};
             memcpy(&m_modbskBarrettMu[i], val, sizeof(DoubleNativeInt));
         }
